Moved manual inpunts to ManualController.cs
