+++ Assets/Scripts/ManualController.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: c6f3154c4d346ae4fb5fa79f61556a2f
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
